% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{basic}
\alias{basic}
\title{basic}
\usage{
basic(data, column)
}
\arguments{
\item{data}{Object type data.frame}

\item{column}{Data column}
}
\value{
data.frame with the basic statistical measures
}
\description{
Call the basicStats function
}
\details{
Call the basicStats function, and plot the data boxplot
}
