% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{freq_dist}
\alias{freq_dist}
\title{freq_dist}
\usage{
freq_dist(data, column, k = NULL)
}
\arguments{
\item{data}{Object type data.frame}

\item{column}{Data variable column}

\item{k}{Number of classes, if null is calculated following the Sturges rule}
}
\value{
data.frame with the frequency distribution
}
\description{
Calculates the frequency distribution
}
\details{
Use the vector data and return the data.frame with the frequency distribution
}
