% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.r
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(file, h, separator, decimal)
}
\arguments{
\item{file}{Data file name}

\item{h}{Boolean if there is a header in the file}

\item{separator}{Column separator used from file}

\item{decimal}{Separator for decimal values in the file}
}
\value{
Object type data.frame
}
\description{
Load database and return a data.frame.
}
\details{
Use the read.table function to load the data and return a table with it.
}
