% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.r
\name{load_geodata}
\alias{load_geodata}
\title{load_geodata}
\usage{
load_geodata(data, coords, column)
}
\arguments{
\item{data}{Object type data.frame}

\item{coords}{Columns of data coordinates in UTM}

\item{column}{Data variable column}
}
\value{
Georeferenced data (class geodata)
}
\description{
Load database and return a data.frame.
}
\details{
Use the data.frame and return Georeferenced data (class geodata) with it.
}
