% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{plot_hist}
\alias{plot_hist}
\title{plot_hist}
\usage{
plot_hist(
  data,
  k,
  fi,
  title = "Histogram",
  x_name = "Data",
  y_name = "Frequency"
)
}
\arguments{
\item{data}{Object type data.frame}

\item{k}{Number of classes}

\item{fi}{Absolute frequency vector}

\item{title}{Histogram title}

\item{x_name}{Label for x axis}

\item{y_name}{Label for y axis}
}
\value{
Plot histogram
}
\description{
Plot histogram
}
\details{
Plots the histogram with vector data, frequency and number of classes
}
