% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.r
\name{setdir}
\alias{setdir}
\title{setdir}
\usage{
setdir(dir = "")
}
\arguments{
\item{dir}{Project root directory}
}
\description{
Change the working directory
}
\details{
Change the working directory to the specified one
}
