% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{statistical_analysis}
\alias{statistical_analysis}
\title{statistical_analysis}
\usage{
statistical_analysis(data, column)
}
\arguments{
\item{data}{Object type data.frame}

\item{column}{Data variable column}
}
\value{
data frame with the results
}
\description{
Performs normality tests
}
\details{
Perform normality tests: Kolmogorov-Smirnov and Shapiro-Wilk test
}
